+++ rust/main/agents/relayer/src/relayer.rs
@@ -22,10 +22,10 @@ use tracing::{error, info, info_span, instrument::Instrumented, warn, Instrument
 use hyperlane_base::{
     broadcast::BroadcastMpscSender,
     db::{HyperlaneRocksDB, DB},
-    metrics::{AgentMetrics, MetricsUpdater},
+    metrics::{AgentMetrics, ChainSpecificMetricsUpdater},
     settings::{ChainConf, IndexSettings},
     AgentMetadata, BaseAgent, ChainMetrics, ContractSyncMetrics, ContractSyncer, CoreMetrics,
-    HyperlaneAgentCore, SyncOptions,
+    HyperlaneAgentCore, RuntimeMetrics, SyncOptions,
 };
 use hyperlane_core::{
     rpc_clients::call_and_retry_n_times, ChainCommunicationError, ContractSyncCursor,
@@ -91,6 +91,7 @@ pub struct Relayer {
     // or move them in `core_metrics`, like the validator metrics
     agent_metrics: AgentMetrics,
     chain_metrics: ChainMetrics,
+    runtime_metrics: RuntimeMetrics,
     /// Tokio console server
     pub tokio_console_server: Option<console_subscriber::Server>,
 }
@@ -125,6 +126,7 @@ impl BaseAgent for Relayer {
         core_metrics: Arc<CoreMetrics>,
         agent_metrics: AgentMetrics,
         chain_metrics: ChainMetrics,
+        runtime_metrics: RuntimeMetrics,
         tokio_console_server: console_subscriber::Server,
     ) -> Result<Self>
     where
@@ -311,6 +313,7 @@ impl BaseAgent for Relayer {
             core_metrics,
             agent_metrics,
             chain_metrics,
+            runtime_metrics,
             tokio_console_server: Some(tokio_console_server),
         })
     }
@@ -321,13 +324,15 @@ impl BaseAgent for Relayer {
 
         let task_monitor = tokio_metrics::TaskMonitor::new();
         if let Some(tokio_console_server) = self.tokio_console_server.take() {
-            let console_server =
-                tokio::spawn(TaskMonitor::instrument(&task_monitor.clone(), async move {
+            let console_server = tokio::task::Builder::new()
+                .name("tokio_console_server")
+                .spawn(TaskMonitor::instrument(&task_monitor.clone(), async move {
                     info!("Starting tokio console server");
                     if let Err(e) = tokio_console_server.serve().await {
                         error!(error=?e, "Tokio console server failed to start");
                     }
-                }));
+                }))
+                .expect("spawning tokio task from Builder is infallible");
             tasks.push(console_server.instrument(info_span!("Tokio console server")));
         }
         let sender = BroadcastSender::new(ENDPOINT_MESSAGES_QUEUE_SIZE);
@@ -358,7 +363,7 @@ impl BaseAgent for Relayer {
                 task_monitor.clone(),
             ));
 
-            let metrics_updater = MetricsUpdater::new(
+            let metrics_updater = ChainSpecificMetricsUpdater::new(
                 dest_conf,
                 self.core_metrics.clone(),
                 self.agent_metrics.clone(),
@@ -388,7 +393,7 @@ impl BaseAgent for Relayer {
                 .await,
             );
             tasks.push(
-                self.run_merkle_tree_hook_syncs(
+                self.run_merkle_tree_hook_sync(
                     origin,
                     BroadcastMpscSender::map_get_receiver(maybe_broadcaster.as_ref()).await,
                     task_monitor.clone(),
@@ -421,6 +426,8 @@ impl BaseAgent for Relayer {
             ));
             tasks.push(self.run_merkle_tree_processor(origin, task_monitor.clone()));
         }
+
+        tasks.push(self.runtime_metrics.spawn());
 
         if let Err(err) = try_join_all(tasks).await {
             tracing::error!(
@@ -478,12 +485,16 @@ impl Relayer {
             }
         };
         let origin_name = origin.name().to_string();
-        tokio::spawn(TaskMonitor::instrument(&task_monitor, async move {
-            let label = "dispatched_messages";
-            contract_sync.clone().sync(label, cursor.into()).await;
-            info!(chain = origin_name, label, "contract sync task exit");
-        }))
-        .instrument(info_span!("MessageSync"))
+        let name = Self::contract_sync_task_name("message::", &origin_name);
+        tokio::task::Builder::new()
+            .name(&name)
+            .spawn(TaskMonitor::instrument(&task_monitor, async move {
+                let label = "dispatched_messages";
+                contract_sync.clone().sync(label, cursor.into()).await;
+                info!(chain = origin_name, label, "contract sync task exit");
+            }))
+            .expect("spawning tokio task from Builder is infallible")
+            .instrument(info_span!("MessageSync"))
     }
 
     async fn run_interchain_gas_payment_sync(
@@ -509,18 +520,22 @@ impl Relayer {
             }
         };
         let origin_name = origin.name().to_string();
-        tokio::spawn(TaskMonitor::instrument(&task_monitor, async move {
-            let label = "gas_payments";
-            contract_sync
-                .clone()
-                .sync(label, SyncOptions::new(Some(cursor), tx_id_receiver))
-                .await;
-            info!(chain = origin_name, label, "contract sync task exit");
-        }))
-        .instrument(info_span!("IgpSync"))
+        let name = Self::contract_sync_task_name("gas_payment::", &origin_name);
+        tokio::task::Builder::new()
+            .name(&name)
+            .spawn(TaskMonitor::instrument(&task_monitor, async move {
+                let label = "gas_payments";
+                contract_sync
+                    .clone()
+                    .sync(label, SyncOptions::new(Some(cursor), tx_id_receiver))
+                    .await;
+                info!(chain = origin_name, label, "contract sync task exit");
+            }))
+            .expect("spawning tokio task from Builder is infallible")
+            .instrument(info_span!("IgpSync"))
     }
 
-    async fn run_merkle_tree_hook_syncs(
+    async fn run_merkle_tree_hook_sync(
         &self,
         origin: &HyperlaneDomain,
         tx_id_receiver: Option<MpscReceiver<H512>>,
@@ -539,15 +554,23 @@ impl Relayer {
             }
         };
         let origin_name = origin.name().to_string();
-        tokio::spawn(TaskMonitor::instrument(&task_monitor, async move {
-            let label = "merkle_tree_hook";
-            contract_sync
-                .clone()
-                .sync(label, SyncOptions::new(Some(cursor), tx_id_receiver))
-                .await;
-            info!(chain = origin_name, label, "contract sync task exit");
-        }))
-        .instrument(info_span!("MerkleTreeHookSync"))
+        let name = Self::contract_sync_task_name("merkle_tree::", &origin_name);
+        tokio::task::Builder::new()
+            .name(&name)
+            .spawn(TaskMonitor::instrument(&task_monitor, async move {
+                let label = "merkle_tree_hook";
+                contract_sync
+                    .clone()
+                    .sync(label, SyncOptions::new(Some(cursor), tx_id_receiver))
+                    .await;
+                info!(chain = origin_name, label, "contract sync task exit");
+            }))
+            .expect("spawning tokio task from Builder is infallible")
+            .instrument(info_span!("MerkleTreeHookSync"))
+    }
+
+    fn contract_sync_task_name(prefix: &str, domain: &str) -> String {
+        format!("contract::sync::{}{}", prefix, domain)
     }
 
     fn run_message_processor(
@@ -621,16 +644,20 @@ impl Relayer {
     ) -> Instrumented<JoinHandle<()>> {
         let span = info_span!("SerialSubmitter", destination=%destination);
         let destination = destination.clone();
-        tokio::spawn(TaskMonitor::instrument(&task_monitor, async move {
-            // Propagate task panics
-            serial_submitter.spawn().await.unwrap_or_else(|err| {
-                panic!(
-                    "destination submitter panicked for destination {}: {:?}",
-                    destination, err
-                )
-            });
-        }))
-        .instrument(span)
+        let name = format!("submitter::destination::{}", destination.name());
+        tokio::task::Builder::new()
+            .name(&name)
+            .spawn(TaskMonitor::instrument(&task_monitor, async move {
+                // Propagate task panics
+                serial_submitter.spawn().await.unwrap_or_else(|err| {
+                    panic!(
+                        "destination submitter panicked for destination {}: {:?}",
+                        destination, err
+                    )
+                });
+            }))
+            .expect("spawning tokio task from Builder is infallible")
+            .instrument(span)
     }
 
     /// Helper function to build and return a hashmap of mailboxes.