+++ rust/main/chains/hyperlane-sealevel/src/application/operation_verifier.rs
@@ -20,7 +20,7 @@ use crate::SealevelProvider;
 
 const NATIVE_WARP_ROUTE_PREFIX: &str = "SOL/";
 // Native SOL warp routers
-// from https://github.com/hyperlane-xyz/hyperlane-registry/tree/359bd01a99cf3c9f2f938c4bc200c5261e2f5bc1/deployments/warp_routes/SOL
+// from https://github.com/hyperlane-xyz/hyperlane-registry/tree/7f605c771cc2ef1fb9f2200605c0885bb453c6d3/deployments/warp_routes/SOL
 lazy_static! {
     static ref NATIVE_WARP_ROUTES: HashSet<H256> = {
         HashSet::from([
@@ -28,6 +28,7 @@ lazy_static! {
             hex_or_base58_to_h256("BXKDfnNkgUNVT5uCfk36sv2GDtK6RwAt9SLbGiKzZkih").unwrap(),
             hex_or_base58_to_h256("7KD647mgysBeEt6PSrv2XYktkSNLzear124oaMENp8SY").unwrap(),
             hex_or_base58_to_h256("GPFwRQ5Cw6dTWnmappUKJt76DD8yawxPx28QugfCaGaA").unwrap(),
+            hex_or_base58_to_h256("4CMbJtieJ7EboZZGSbXTQjW5i2sL638jFvE3dWTYG3SK").unwrap(),
         ])
     };
 }