+++ rust/main/chains/hyperlane-sealevel/src/interchain_security_module.rs
@@ -10,7 +10,7 @@ use hyperlane_core::{
 use hyperlane_sealevel_interchain_security_module_interface::InterchainSecurityModuleInstruction;
 use serializable_account_meta::SimulationReturnData;
 
-use crate::{ConnectionConf, SealevelKeypair, SealevelProvider, SealevelRpcClient};
+use crate::{SealevelKeypair, SealevelProvider, SealevelRpcClient};
 
 /// A reference to an InterchainSecurityModule contract on some Sealevel chain
 #[derive(Debug)]
@@ -23,11 +23,10 @@ pub struct SealevelInterchainSecurityModule {
 impl SealevelInterchainSecurityModule {
     /// Create a new sealevel InterchainSecurityModule
     pub fn new(
-        conf: &ConnectionConf,
+        provider: SealevelProvider,
         locator: ContractLocator,
         payer: Option<SealevelKeypair>,
     ) -> Self {
-        let provider = SealevelProvider::new(locator.domain.clone(), conf);
         let program_id = Pubkey::from(<[u8; 32]>::from(locator.address));
         Self {
             payer,